Moves Page component to shared
