+++ src/templates/category-template.js
@@ -4,7 +4,7 @@ import { Row, Col } from 'react-grid-system';
 import PropTypes from 'prop-types';
 import Layout from '../components/shared/Layout';
 import Courses from '../components/Courses';
-import Page from '../components/Page';
+import Page from '../components/shared/Page';
 import Pagination from '../components/Pagination';
 import { PAGE_CONTEXT_PROP_TYPE, SITE_METADATA_PROP_TYPE } from '../constants/propTypes';
