+++ src/templates/course-list-template.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import Layout from '../components/shared/Layout';
 import Courses from '../components/Courses';
-import Page from '../components/Page';
+import Page from '../components/shared/Page';
 import Pagination from '../components/Pagination';
 
 const CourseListTemplate = ({ data, pageContext }) => {
