+++ src/templates/page-template.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import { Container, Row, Col } from 'react-grid-system';
 import Layout from '../components/shared/Layout';
-import Page from '../components/Page';
+import Page from '../components/shared/Page';
 
 const PageTemplate = ({ data }) => {
   const { title: siteTitle, subtitle: siteSubtitle } = data.site.siteMetadata;
