+++ src/templates/tag-template.js
@@ -4,7 +4,7 @@ import { Container, Row, Col } from 'react-grid-system';
 import PropTypes from 'prop-types';
 import Layout from '../components/shared/Layout';
 import Courses from '../components/Courses';
-import Page from '../components/Page';
+import Page from '../components/shared/Page';
 import Pagination from '../components/Pagination';
 import { SITE_METADATA_PROP_TYPE } from '../constants/propTypes';
