+++ src/templates/tags-list-template.js
@@ -3,7 +3,7 @@ import { Link, graphql } from 'gatsby';
 import kebabCase from 'lodash/kebabCase';
 import { Container, Row, Col } from 'react-grid-system';
 import Layout from '../components/shared/Layout';
-import Page from '../components/Page';
+import Page from '../components/shared/Page';
 
 const TagsListTemplate = ({ data }) => {
   const { title, subtitle } = data.site.siteMetadata;
